+++ packages/shared/src/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from './roles';